+++ app/src/main/res/layout/maps_activity.xml
@@ -1,7 +1,19 @@
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/map"
-    android:name="com.google.android.gms.maps.SupportMapFragment"
+    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.mapwithmarker.MapsMarkerActivity" />
+    android:layout_height="match_parent">
+
+
+
+    <fragment
+        android:layout_weight="1"
+        android:id="@+id/map"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="0dp" />
+    <com.github.mikephil.charting.charts.LineChart
+        android:id="@+id/ride_detail_speed_chart"
+        android:layout_width="match_parent"
+        android:layout_height="150dp" />
+</LinearLayout>
